// cmp_mini_frontend/src/servers/models/GoodsVO.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type GoodsVO = {
    categoryId?: number;
    createTime?: string;
    description?: string;
    id?: number;
    image?: string;
    name?: string;
    price?: number;
    quantity?: number;
    stock?: number;
    updateTime?: string;
};